refactor: showActivityViewController 메서드와 showDeleteActionAlert 메서드를 기본구현하는 DiaryPresentable 프로토콜 생성 및 채택
